polyhedral model: Don't store array & stmt pointer as domain ID data
That would effectively change the domain ID and hence the domain,
when an array or statement is constructed.

That can cause confusion when the domain passed to constructor is
further used to create dependence relations etc., because the domain
is not considered identical to that of array or statement anymore.

Fortunately, the pointer stored in ID is not needed anymore,
so simply avoid storing it.